Refactor ConnectionAdapterPatch
