--- lib/arproxy/connection_adapter_patches/trilogy_patch.rb
@@ -1,11 +1,0 @@
-require 'arproxy/connection_adapter_patches/base_patch'
-
-module Arproxy
-  module ConnectionAdapterPatches
-    class TrilogyPatch < BasePatch
-      def enable!
-        enable_patches :raw_execute
-      end
-    end
-  end
-end
